Feat: 게시글 상세 조회(수정용) 기능 추가 및 ID 타입 Long으로 수정
- 게시글 수정 전용 상세조회 엔드포인트 추가

- 기존 DTO, Mapper, Service 중 id Integer -> Long 타입 변경

+++ src/main/java/com/example/mockvoting/domain/community/dto/PostDetailResponseDTO.java
@@ -13,8 +13,8 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PostDetailResponseDTO {
-    private Integer id;
-    private Integer categoryId;
+    private Long id;
+    private Long categoryId;
     private String title;
     private String content;
     private String authorId;
